Make LibK::vector<T> more spec-conform; bump C++ standard to C++23
- Started work on improving standard compliance for the kernel STL
  implementation
- Decided to target C++23 STL (for now without use of concepts)
- Implemented std::allocator as LibK::allocator
	- stateless allocator with is_always_equal and propagation on
	  move, copy and swap
	- uses global new and delete operators
- removed deprecated std::iterator argument for LibK::normal_iterator
- added conversion operator to const_iterator
- made LibK::vector<T> mostly spec-conform with std::vector<T>
	- it now uses an allocator for memory management
	- every method is constexpr as mandated by spec
	- not sure about move and copy construction with allocators
	- not sure wheter allocation up to m_capacity is *correct* as it
	  constructs objects in unused memory (at least I think it does)
- rewrote LibK::basic_string<charT> as a wrapper around
  LibK::vector<charT>
- had to define max_align_t in some cases (it should be defined by
  <cstddef> but it only defines it in some instances somehow; may be due
  to my incomplete libc headers the hosted toolchain references)
- fixed various bugs caused by switching to the hosted toolchain and to
  C++23
  	- added casts and template arguments in cases where C++23 is
	  stricter than C++20
	- global constructors where moved to (.init_array)
	- stack protection required __stack_chk_fail_local()

+++ src/kernel/storage/ata/AHCIPort.cpp
@@ -183,9 +183,9 @@ namespace Kernel
 		return processed_count;
 	}
 
-	size_t AHCIPort::find_slot()
+	uint32_t AHCIPort::find_slot()
 	{
-		for (size_t i = 0; i < AHCI::NUM_SLOTS; i++)
+		for (uint32_t i = 0; i < AHCI::NUM_SLOTS; i++)
 		{
 			if (!(m_hba_port->ci & (1 << i)) && !m_command_slots[i].issued())
 				return i;
